requirements.txt
aiohttp==3.9.5
aioimaplib==1.1.0
aiosignal==1.3.1
aiosmtplib==3.0.1
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
DateTime==5.5
distro==1.9.0
frozenlist==1.4.1
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
gspread==6.1.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
multidict==6.0.5
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.35.1
proto-plus==1.24.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
pyparsing==3.1.2
python-dotenv==1.0.1
pytz==2024.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==70.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.9.4
zope.interface==6.4.post2
